validate UTF8-ness in Delete, Attributes, SignedURL, NewRangeReader

+++ blob/blob.go
@@ -540,6 +540,10 @@ func (b *Bucket) List(opts *ListOptions) *ListIterator {
 // If the blob does not exist, Attributes returns an error for which
 // gcerrors.Code will return gcerrors.NotFound.
 func (b *Bucket) Attributes(ctx context.Context, key string) (_ Attributes, err error) {
+	if !utf8.ValidString(key) {
+		return Attributes{}, fmt.Errorf("blob.Attributes: key must be a valid UTF-8 string: %q", key)
+	}
+
 	ctx = b.tracer.Start(ctx, "Attributes")
 	defer func() { b.tracer.End(ctx, err) }()
 
@@ -590,6 +594,9 @@ func (b *Bucket) NewReader(ctx context.Context, key string, opts *ReaderOptions)
 func (b *Bucket) NewRangeReader(ctx context.Context, key string, offset, length int64, opts *ReaderOptions) (_ *Reader, err error) {
 	if offset < 0 {
 		return nil, errors.New("blob.NewRangeReader: offset must be non-negative")
+	}
+	if !utf8.ValidString(key) {
+		return nil, fmt.Errorf("blob.NewRangeReader: key must be a valid UTF-8 string: %q", key)
 	}
 	if opts == nil {
 		opts = &ReaderOptions{}
@@ -732,6 +739,9 @@ func (b *Bucket) NewWriter(ctx context.Context, key string, opts *WriterOptions)
 // If the blob does not exist, Delete returns an error for which
 // gcerrors.Code will return gcerrors.NotFound.
 func (b *Bucket) Delete(ctx context.Context, key string) (err error) {
+	if !utf8.ValidString(key) {
+		return fmt.Errorf("blob.Delete: key must be a valid UTF-8 string: %q", key)
+	}
 	ctx = b.tracer.Start(ctx, "Delete")
 	defer func() { b.tracer.End(ctx, err) }()
 	return wrapError(b.b, b.b.Delete(ctx, key))
@@ -747,6 +757,9 @@ func (b *Bucket) Delete(ctx context.Context, key string) (err error) {
 // If the provider implementation does not support this functionality, SignedURL
 // will return an error for which gcerrors.Code will return gcerrors.Unimplemented.
 func (b *Bucket) SignedURL(ctx context.Context, key string, opts *SignedURLOptions) (string, error) {
+	if !utf8.ValidString(key) {
+		return "", fmt.Errorf("blob.SignedURL: key must be a valid UTF-8 string: %q", key)
+	}
 	if opts == nil {
 		opts = &SignedURLOptions{}
 	}
